Added Label Position
git-svn-id: svn+ssh://erebor.esa.informatik.tu-darmstadt.de/home/wimi/svn/GaalopCompiler/gaalop-2.0/branches/tbaAndGapp@1234 4235ee25-d3ef-403b-b6ed-692c26853d4b

+++ codegen-vis2d/src/main/java/de/gaalop/vis2d/drawing/Text2d.java
@@ -8,11 +8,11 @@ import java.awt.Color;
  */
 public class Text2d extends DrawObject {
     
-    public int x;
-    public int y;
+    public double x;
+    public double y;
     public String text;
 
-    public Text2d(int x, int y, String text, Color color) {
+    public Text2d(double x, double y, String text, Color color) {
         super(color);
         this.x = x;
         this.y = y;
